+++ cli/srecli/group/metadata.py
@@ -1,8 +1,11 @@
 import click
 
 from srecli.group import (
-    ContextObj, ViewCommand, cli_response
+    ContextObj, ViewCommand, cli_response,
+    build_background_job_status_command
 )
+from srecli.service.adapter_client import SREResponse
+from srecli.service.constants import BackgroundJobName
 
 
 @click.group(name='metadata')
@@ -10,9 +13,15 @@ def metadata():
     """Manages locally stored metadata"""
 
 
-@metadata.command(cls=ViewCommand, name='update')
+@metadata.command(
+    cls=ViewCommand, 
+    name='update',
+)
 @cli_response()
-def update(ctx: ContextObj, customer_id: str | None = None):
+def update(
+    ctx: ContextObj,
+    customer_id: str | None = None,
+) -> SREResponse:
     """
     Triggers a complete synchronization with License Manager for all customers.
     
@@ -24,3 +33,9 @@ def update(ctx: ContextObj, customer_id: str | None = None):
     """
     return ctx['api_client'].trigger_metadata_update()
 
+
+build_background_job_status_command(
+    group=metadata,
+    background_job_name=BackgroundJobName.METADATA,
+    help_text='Execution status of the last metadata update',
+)