+++ cli/srecli/group/metrics.py
@@ -2,8 +2,10 @@ import click
 
 from srecli.group import (
     ContextObj, ViewCommand, cli_response,
-    DYNAMIC_DATE_ONLY_EXAMPLE, DYNAMIC_DATE_ONLY_PAST_EXAMPLE
+    build_background_job_status_command
 )
+from srecli.service.adapter_client import SREResponse
+from srecli.service.constants import BackgroundJobName
 
 
 @click.group(name='metrics')
@@ -11,9 +13,15 @@ def metrics():
     """Manages Scan and Tenant Metrics"""
 
 
-@metrics.command(cls=ViewCommand, name='update')
+@metrics.command(
+    cls=ViewCommand, 
+    name='update',
+)
 @cli_response()
-def update(ctx: ContextObj, customer_id):
+def update(
+    ctx: ContextObj,
+    customer_id: str | None = None,
+) -> SREResponse:
     """
     Triggers a metrics update for Syndicate Rule Engine reports. Report data will
     contain data up to the time when the trigger was executed
@@ -21,20 +29,8 @@ def update(ctx: ContextObj, customer_id):
     return ctx['api_client'].trigger_metrics_update()
 
 
-@metrics.command(cls=ViewCommand, name='status')
-@click.option('--from_date', '-from', type=str,
-              help='Query metrics statuses from this date. Accepts date ISO '
-                   f'string. Example: {DYNAMIC_DATE_ONLY_PAST_EXAMPLE}')
-@click.option('--to_date', '-to', type=str,
-              help='Query metrics statuses till this date. Accepts date ISO '
-                   f'string. Example: {DYNAMIC_DATE_ONLY_EXAMPLE}')
-@cli_response()
-def status(ctx: ContextObj, from_date: str, to_date: str, customer_id):
-    """
-    Execution status of the last metrics update
-    """
-    params = {
-        'from': from_date,
-        'to': to_date
-    }
-    return ctx['api_client'].metrics_status(**params)
+build_background_job_status_command(
+    group=metrics,
+    background_job_name=BackgroundJobName.METRICS,
+    help_text='Execution status of the last metrics update',
+)