+++ src/lambdas/api_handler/handler.py
@@ -13,8 +13,8 @@ from lambdas.api_handler.handlers.health_check_handler import (
     HealthCheckHandler,
 )
 from lambdas.api_handler.handlers.job_handler import JobHandler
-from lambdas.api_handler.handlers.metrics_status_handler import (
-    MetricsStatusHandler,
+from lambdas.api_handler.handlers.background_job_status_handler import (
+    BackgroundJobStatusHandler,
 )
 from lambdas.api_handler.handlers.new_swagger_handler import \
     SwaggerHandler
@@ -47,9 +47,9 @@ class ApiHandler(ApiEventProcessorLambdaHandler):
             UsersHandler,
             JobHandler,
             HealthCheckHandler,
-            MetricsStatusHandler,
             EventsHandler,
             SwaggerHandler,
+            BackgroundJobStatusHandler,
         ]
 
     @classmethod